+++ app/src/main/java/nbc/group/recipes/presentation/BookMarkAdapter.kt
@@ -1,12 +1,11 @@
-package nbc.group.recipes
+package nbc.group.recipes.presentation
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
-import nbc.group.recipes.data.model.dto.Item
+import nbc.group.recipes.R
 import nbc.group.recipes.data.model.dto.Recipe
 import nbc.group.recipes.databinding.BookmarkItemBinding
