+++ app/src/main/java/nbc/group/recipes/presentation/fragment/BookmarkFragment.kt
@@ -14,8 +14,8 @@ import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
-import nbc.group.recipes.BookMarkAdapter
-import nbc.group.recipes.BookMarkViewModel
+import nbc.group.recipes.presentation.BookMarkAdapter
+import nbc.group.recipes.viewmodel.BookMarkViewModel
 import nbc.group.recipes.VisibilityView
 import nbc.group.recipes.data.model.dto.Recipe
 import nbc.group.recipes.data.model.dto.toRecipe
