+++ app/src/main/java/nbc/group/recipes/presentation/fragment/BottomSheetFragment.kt
@@ -1,6 +1,5 @@
-package nbc.group.recipes
+package nbc.group.recipes.presentation.fragment
 
-import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
@@ -15,6 +14,7 @@ import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
+import nbc.group.recipes.presentation.BottomSheetAdapter
 import nbc.group.recipes.databinding.FragmentBottomsheetBinding
 import nbc.group.recipes.presentation.MainActivity
 import nbc.group.recipes.viewmodel.MapViewModel
