+++ app/src/main/java/nbc/group/recipes/viewmodel/BookMarkViewModel.kt
@@ -1,4 +1,4 @@
-package nbc.group.recipes
+package nbc.group.recipes.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -7,6 +7,7 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
+import nbc.group.recipes.VisibilityView
 import nbc.group.recipes.data.local.dao.RecipeDao
 import nbc.group.recipes.data.model.entity.RecipeEntity
 import javax.inject.Inject
